+++ Cross.CQRS.EF/Options/TransactionBehaviorOptions.cs
@@ -1,7 +1,7 @@
-namespace Cross.CQRS.EF.Options;
+﻿namespace Cross.CQRS.EF.Options;
 
 public class TransactionBehaviorOptions
 {
-    public TransactionBehaviorEnum Behavior { get; set; } = TransactionBehaviorEnum.TransactionalScopeBehavior;
-    public IsolationLevel IsolationLevel { get; set; } = IsolationLevel.ReadCommitted;
+    public TransactionBehaviorEnum Behavior { get; set; } = TransactionBehaviorEnum.TransactionalBehavior;
+    public IsolationLevel IsolationLevel { get; set; } = IsolationLevel.Serializable;
 }